feat(overlay): move some overlay directive stuff to the map to allow for more flexibility when overlaying stuf
