+++ src/lib/layer/shared/layers/vector-layer.interface.ts
@@ -2,5 +2,5 @@ import { LayerOptions } from './layer.interface';
 
 export interface VectorLayerOptions extends LayerOptions {
   view?: olx.layer.VectorOptions;
-  style?: olx.style.StyleOptions;
+  style?: any;
 }
